Rename contact us page sections file
